refactor: extract shared securePath seam from path-sandbox callers
Two adapters were drifting: isPathWithinRoot was duplicated verbatim in
normalizePdfInput.ts (allowedInputRoot sandbox) and diffOutputGuards.ts
(diffsOutputFolder sandbox), and the symlink-rejection walker existed
only inside diffOutputGuards.ts. A third sandbox callsite for
pdfToPngConvertOptions.outputFolder is about to land — consolidating the
CWE-59 / CWE-61 symlink-swap defense in one module is the prerequisite
that keeps the fix in a single place.

The new src/internal/securePath.ts exports:
- isPathWithinRoot(pathToCheck, rootPath)
- pathExistsWithoutFollowingSymlinks(path)
- assertPathIsNotSymbolicLink(path, errorMessage)
- assertPathAndAncestorsAreNotSymbolicLinks(path, errorMessage)

Both prior callers now import from the seam. The symlink-rejection
helpers accept an explicit error message so callers retain their domain
language (\"Diff output path must stay within diffsOutputFolder: ...\")
while sharing one walk implementation. The thrown error class is
ComparePdfConfigurationError for every caller, matching prior behavior.

No behavior changes — covered by the existing 161-test suite (symlink
TOCTOU tests in __tests__/18.*.test.ts continue to pass).
